feat : load all items to frontends

+++ internal/Items/items.go
@@ -9,6 +9,6 @@ type Item struct {
 	Health      int16  `json:"health"`
 	Mood        int16  `json:"mood"`
 	Energy      int16  `json:"energy"`
-	MoneyCost   int16  `json:"moneyCost"`
+	MoneyCost   int  `json:"moneyCost"`
 	Description string `json:"description"`
 }
